apm0013173-css-ra-velocloud/requirements.txt
requests
Django==3.2.3
gunicorn
djangorestframework==3.11.1
drf-yasg
coverage
xlsxwriter
jsonpath-ng
websocket-client
numpy
pandas
ipaddress